AirSense/terminal_website/requirements.txt
flask
flask-cors
pandas
requests
google-genai
google-adk
matplotlib

